requirements.txt
streamlit>=1.28.0
tensorflow>=2.13.0
tensorflow_hub>=0.14.0
pillow>=9.0.0
numpy>=1.21.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.15.0
opencv-python>=4.5.0
scikit-learn>=1.1.0
